[rust] Clarify template build files
Rename the Cargo.toml and lib.rs from the template crate to have a `.in`
suffix. This makes it clearer that these aren't just the files used, but
are templates. Currently, there are no changes made, and they are just
copied, but the build comes from the copies.

These files, especially `lib.rs.in` should remain minimal, as they will
be awkward to edit, not having rust-analyzer support.

Signed-off-by: David Brown <david.brown@linaro.org>