rust: Add 'unsafe' PL011 driver
A fairly mechanical translation of the existing C driver to Rust. This
is functionally equivalent, and uses direct unsafe memory access in a
similar manner to the C code.

This use Rust's volatile for the register access, which works on this
platform which doesn't have any barriers for IO. But that mechanism does
need to be dealt with properly in Rust.

Signed-off-by: David Brown <david.brown@linaro.org>