rust: dev-pl011: Fix some PL011 register names
Remove an extraneous 't' from the beginning of some of the PL011
register names. The names in the docs are in all caps, and have no space
between the "UART" prefix and the rest of the name. That 't' leaked into
three of the register names. Remove this so the names actually match
their descriptions.

Signed-off-by: David Brown <david.brown@linaro.org>