rust: Only use opt-level 2 for dev
Opt level will be a tradeoff as to how debuggable the resulting
executable is. Not optimizing with Rust tends to produce code that is
significantly slower, so allow for some degree of optimization of the
rust code. The previous "z" is rather aggressive, and makes the code
much more difficult to debug.

Signed-off-by: David Brown <david.brown@linaro.org>